requirements.txt
httpx>=0.27.0
pytest>=8.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
google-genai>=1.0.0
